chore: restructure password schemas
