+++ web_application/src/actions/divisions/update.schema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { idSchema } from '../base/get.schema';
+import { idSchema } from '../base/base.schema';
 import {
     divisionAttributesSchema,
     divisionRelationshipsSchema,
