+++ web_application/src/actions/financeContacts/update.schema.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { idSchema } from '../base/get.schema';
+import { idSchema } from '../base/base.schema';
 import { financeContactAttributesSchema } from './create.schema';
 
 export const genderOptions = ['other', 'male', 'female'] as const;
