+++ web_application/src/actions/users/forgotPassword.schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const forgotPasswordSchema = z.object({
+    email: z.email(),
+});
